rolling bank support
RC candidate

This now means that slow sd cards are supported by buffering the entire
download into memory before dropping down to disk.

It's fast enough to keep up with the draining ESP buffer but does impose
limits on the size we can download.

There is also the `-r` flag to disable rolling banks - but this isn't
fully tested yet.